refactor: Consolidate ILibrary interfaces and update tests
- Consolidated multiple interfaces into ILibrary: ILibraryBookAdder,  ILibraryBookRemover, ILibraryBookSorter, ILibraryShower, and  ILibraryBookLocator
- Removed the now redundant interfaces: ILibraryBookAdder,  ILibraryBookRemover, ILibraryBookSorter, ILibraryShower, and  ILibraryBookLocator
- Consolidated test cases into a new ILibraryTests class, removing  ILibraryAddBookTests, ILibraryRemoveBookTests, LibrarySearcherTests,  and ILibraryShowerTests
- Updated file path in XmlBookListLoaderTests from "testLibrary.xml" to  "testLibraryLoad.xml"

refactor: Consolidate multiple interfaces into ILibrary

- Refactored `ILibrary` to combine `ILibraryBookAdder`, `ILibraryBookRemover`,  `ILibraryBookSorter`, `ILibraryShower`, and `ILibraryBookLocator` into one  interface.
- Added methods to `ILibrary` for adding, removing, sorting, displaying, and  finding books, as well as retrieving the total number of books.
- Removed individual interfaces as their functionalities are now part of  `ILibrary`.
- Removed test classes for individual interfaces.
- Added `ILibraryTests` to test the consolidated `ILibrary` interface.
- Updated file path in `XmlBookListLoaderTests` from "testLibrary.xml" to "testLibraryLoad.xml".
