--- BookLibraryManager.Common/ILibraryBookSorter.cs
@@ -1,13 +1,0 @@
-﻿namespace BookLibraryManager.Common;
-
-/// <author>YR 2025-01-09</author>
-/// <summary>
-/// Interface for sorting books in the library.
-/// </summary>
-public interface ILibraryBookSorter
-{
-    /// <summary>
-    /// Sorts Books in the library.
-    /// </summary>
-    void SortLibrary();
-}
